refactor: separate helper functions and fix tests
