+++ packages/plugin/src/__tests__/ChartPlugin.spec.js
@@ -35,18 +35,18 @@ const ouMock = {
     ],
 }
 
-const defaultCurrentMock = {
-    type: analytics.VIS_TYPE_COLUMN,
-    columns: [dxMock],
-    rows: [peMock],
-    filters: [ouMock],
-}
-
 const yearOverYearCurrentMock = {
     type: analytics.VIS_TYPE_YEAR_OVER_YEAR_LINE,
     columns: [dxMock],
     rows: [peMock],
     yearlySeries: ['LAST_YEAR'],
+}
+
+const mockExtraOptions = {
+    dashboard: false,
+    noData: {
+        text: 'No data',
+    },
 }
 
 const singleValueCurrentMock = {
@@ -78,24 +78,10 @@ class MockAnalyticsResponse {
     }
 }
 
-const mockYoYSeriesLabels = ['rainbow', 'rarity']
-class MockYoYAnalyticsResponse {
-    constructor() {
-        return {
-            responses: [analyticsResponse],
-            yearlySeriesLabels: mockYoYSeriesLabels,
-        }
-    }
-}
-
 const createVisualizationMock = {
     visualization: {
         getSVGForExport: () => '<svg />',
     },
-}
-
-const isYearOverYearMockResponse = visType => {
-    return visType === analytics.VIS_TYPE_YEAR_OVER_YEAR_LINE
 }
 
 const isSingleValueMockResponse = visType => {
@@ -118,15 +104,11 @@ describe('ChartPlugin', () => {
 
     beforeEach(() => {
         props = {
-            visualization: {},
-            filters: {},
             style: { height: 100 },
             id: 1,
-            d2: {},
-            forDashboard: false,
             onChartGenerated: jest.fn(),
-            onResponsesReceived: jest.fn(),
-            onError: jest.fn(),
+            responses: [],
+            extraOptions: mockExtraOptions,
         }
         chartPlugin = undefined
 
@@ -165,37 +147,6 @@ describe('ChartPlugin', () => {
             })
         })
 
-        it('includes only options that do not have default value in request', done => {
-            props.visualization = {
-                ...defaultCurrentMock,
-                option1: 'def',
-                option2: null,
-            }
-
-            canvas()
-
-            setTimeout(() => {
-                expect(api.apiFetchAnalytics).toHaveBeenCalled()
-                expect(api.apiFetchAnalytics.mock.calls[0][2]).toEqual({
-                    option1: 'def',
-                })
-
-                done()
-            })
-        })
-
-        it('calls onResponsesReceived callback', done => {
-            canvas()
-
-            setTimeout(() => {
-                expect(props.onResponsesReceived).toHaveBeenCalled()
-                expect(props.onResponsesReceived).toHaveBeenCalledWith([
-                    analyticsResponse,
-                ])
-                done()
-            })
-        })
-
         it('calls onChartGenerated callback', done => {
             canvas()
 
@@ -208,92 +159,12 @@ describe('ChartPlugin', () => {
             })
         })
 
-        it('calls onError callback when an exception is thrown', done => {
-            api.apiFetchAnalytics = jest.fn().mockRejectedValue('error')
-
-            canvas()
-
-            setTimeout(() => {
-                expect(props.onError).toHaveBeenCalled()
-                done()
-            })
-        })
-
-        it('sets period when interpretation selected', done => {
-            const period = 'eons ago'
-            props.filters.relativePeriodDate = period
-
-            canvas()
-
-            setTimeout(() => {
-                expect(api.apiFetchAnalytics).toHaveBeenCalled()
-                expect(api.apiFetchAnalytics.mock.calls[0][2]).toHaveProperty(
-                    'relativePeriodDate',
-                    period
-                )
-
-                done()
-            })
-        })
-
         describe('Year-on-year chart', () => {
             beforeEach(() => {
                 props.visualization = {
                     ...yearOverYearCurrentMock,
                     option1: 'def',
                 }
-
-                api.apiFetchAnalyticsForYearOverYear = jest
-                    .fn()
-                    .mockResolvedValue(new MockYoYAnalyticsResponse())
-
-                analytics.isYearOverYear = jest
-                    .fn()
-                    .mockReturnValue(
-                        isYearOverYearMockResponse(props.visualization.type)
-                    )
-            })
-
-            it('makes year-on-year analytics request', done => {
-                canvas()
-
-                setTimeout(() => {
-                    expect(
-                        api.apiFetchAnalyticsForYearOverYear
-                    ).toHaveBeenCalled()
-                    expect(
-                        api.apiFetchAnalyticsForYearOverYear.mock.calls[0][1]
-                    ).toEqual({
-                        ...yearOverYearCurrentMock,
-                        option1: 'def',
-                    })
-
-                    done()
-                })
-            })
-
-            it('provides extra options to createVisualization', done => {
-                canvas()
-
-                setTimeout(() => {
-                    expect(analytics.createVisualization).toHaveBeenCalled()
-
-                    const expectedExtraOptions = {
-                        yearlySeries: mockYoYSeriesLabels,
-                        xAxisLabels: ['period 1', 'period 2'],
-                        noData: { text: 'No data' },
-                    }
-
-                    expect(
-                        analytics.createVisualization.mock.calls[0][3]
-                    ).toEqual({
-                        animation: 200, // default prop
-                        dashboard: false,
-                        ...expectedExtraOptions,
-                    })
-
-                    done()
-                })
             })
         })
